requirements.txt
instagrapi>=2.1.0
openai>=1.0.0
apify-client>=2.1.0
python-dotenv>=1.0.0
